Update patch set 1
Patch Set 1:

I know you're doing a bunch of LegacyCamera patches based off of a lot of work that I did to get it working on the X2 to allow it to be built with the Gallery2.apk like the Camera app is at the moment, but Gallery2 also expects com.android.camera.DisableCameraReceiver (at least on the first boot) to disable the camera when the user is running the initial setup wizard. This currently doesn't exist in LegacyCamera, so I pulled it from the standard Camera app and it appears to work fine. See: https://github.com/dragonzkiller/android_packages_apps_LegacyCamera/commit/3117552d29facae2d77fc42a4bd65aaf3a1206c6.

Without this commit, the logcat will show an error during initial setup. It's not fatal, but it should still be resolved like in the linked commit.

Patch-set: 1