Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

@Ricardo: Thank you :)

I still see no point in this. We've already moved the one file that needs to be moved and we don't need to squash everything. What's the point in the repo history then if we're just gonna squash everything?

Patch-set: 4