Update patch set 5
Patch Set 5:

i stand with maintaining authorship and all the other reasoning you gave but i had to do this way since i had to go back to AOSP history due the mess that had been submitted and had to reauthor all the changes done since then. and in that i had to cherry pick Jerpelea's patch to add resolutions and rapmv78's corrections to panorama from the mess that i had removed.


from my point of view i did the best that i could have. others are certainly entitled to their opinions and obviously this is not merged yet. open to discussions and if needed someone can rework this again.

Patch-set: 5