Update patch set 5
Patch Set 5:

Squashing commits is a common practice in source code management. it makes maintaining a patchset much easier, especially when having to pull a patch forward to a new branch.

'anything that allows direct avoidance of mandatory EGL extensions' -- This is considered a 'hack' to make legacy hardware work in a version of Android that does not support the legacy hardware. Rather than having a hack to support a few legacy devices, it would be better if those devices had their own ifdef (or equivalent of ifdef). They are exceptions to the rule.

Patch-set: 5