Update patch set 1
Patch Set 1:

What devices are you using this for. Checking with the rest of the team, we don't use this on any supported devices. That's probably the cause for the lack of review.

Patch-set: 1
Reviewer: Björn Lundén <16@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0